Make the DetsysTs correlation deserialize require a repository
We're getting away from that style of serialization, but the DetsysTs variant requires essentially nothing.
That means EVERYTHING deserializes into that code path and nothing into the Direct format.

DetsysTS always writes out the repository field, so that makes it work.

Another workaround is for DetsysTs to write out the `run` field as a list, which forces moving to the Direct parse.